cart CRUD implementation

+++ audiophile_store/views.py
@@ -3,8 +3,8 @@ from rest_framework.generics import  CreateAPIView, DestroyAPIView
 from rest_framework.response import Response
 from rest_framework.serializers import ValidationError
 from rest_framework.status import HTTP_204_NO_CONTENT, HTTP_404_NOT_FOUND
-from .serializers import ProductSerializer, product_data, CartSerializer
-from .models import Product, Cart
+from .serializers import ProductSerializer, ProductImageSerializer, product_data
+from .models import Product, Product_images
 
 # Create your views here.
 class ProductDataView(APIView):
@@ -15,6 +15,10 @@ class ProductDataView(APIView):
 
         return Response(serializer.data)
 
+class CreateProductImageView(CreateAPIView):
+    serializer_class = ProductImageSerializer
+    queryset = Product_images.objects.all()
+
 
 class CreateProductView(CreateAPIView):
     serializer_class = ProductSerializer
@@ -22,10 +26,28 @@ class CreateProductView(CreateAPIView):
 
 
 class DeleteProductView(APIView):
+    def get (self, request, pk):
+        try:
+            item = Product.objects.get(id=pk)
+            return Response(
+                        {
+                            'id': item.pk,
+                            'name': item.name,
+                            'price': item.price,
+                            'Release_date': item.Release_date,
+                        }
+                    )
+        except  Product.DoesNotExist:
+            return Response({'error': 'Product does not exist in the database'}, status=HTTP_204_NO_CONTENT)
+    
     def delete(self, request ,pk):
-        item = Product.objects.get(id=pk)
-        if not item:
-            return Response(status=HTTP_404_NOT_FOUND)
-        else:
-            item.delete()
-        return Response('Data successfully deleted', status=HTTP_204_NO_CONTENT)
+        try:
+            item = Product.objects.get(id=pk)
+            if request.method == 'DELETE':
+                if not item:
+                    return Response(status=HTTP_404_NOT_FOUND)
+                else:
+                    item.delete()
+            return Response('Data successfully deleted', status=HTTP_204_NO_CONTENT)
+        except  Product.DoesNotExist:
+            return Response({'error': 'Product does not exist in the database'}, status=HTTP_204_NO_CONTENT)
